requirements.txt
asgiref==3.4.1
Django==3.2.6
django-jalali-date==0.3.2
jdatetime==3.6.2
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
mysqlclient==2.1.0